Modified types : phone number and amount

+++ prix_banque_flutter_interface/lib/transfers_management/creation_transfer_page.dart
@@ -130,7 +130,7 @@ class _CreateTransferPage extends State<CreateTransferPage> {
         //mailAdressTransferPayer
         emailReceiverController.text,
         //mailAdressTransferReceiver
-        int.parse(amountController.text),
+        double.parse(amountController.text),
         //transferAmount
         transferType,
         //transferType
